restructured database calling of basic pokemon detail
